+++ fftw/src/main/java/org/bytedeco/javacpp/fftw3.java
@@ -106,13 +106,13 @@ public static class fftw_iodim_do_not_use_me extends Pointer {
     static { Loader.load(); }
     /** Default native constructor. */
     public fftw_iodim_do_not_use_me() { super((Pointer)null); allocate(); }
-    /** Native array allocator. Access with {@link Pointer#position(int)}. */
-    public fftw_iodim_do_not_use_me(int size) { super((Pointer)null); allocateArray(size); }
+    /** Native array allocator. Access with {@link Pointer#position(long)}. */
+    public fftw_iodim_do_not_use_me(long size) { super((Pointer)null); allocateArray(size); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public fftw_iodim_do_not_use_me(Pointer p) { super(p); }
     private native void allocate();
-    private native void allocateArray(int size);
-    @Override public fftw_iodim_do_not_use_me position(int position) {
+    private native void allocateArray(long size);
+    @Override public fftw_iodim_do_not_use_me position(long position) {
         return (fftw_iodim_do_not_use_me)super.position(position);
     }
 
@@ -126,13 +126,13 @@ public static class fftw_iodim64_do_not_use_me extends Pointer {
     static { Loader.load(); }
     /** Default native constructor. */
     public fftw_iodim64_do_not_use_me() { super((Pointer)null); allocate(); }
-    /** Native array allocator. Access with {@link Pointer#position(int)}. */
-    public fftw_iodim64_do_not_use_me(int size) { super((Pointer)null); allocateArray(size); }
+    /** Native array allocator. Access with {@link Pointer#position(long)}. */
+    public fftw_iodim64_do_not_use_me(long size) { super((Pointer)null); allocateArray(size); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public fftw_iodim64_do_not_use_me(Pointer p) { super(p); }
     private native void allocate();
-    private native void allocateArray(int size);
-    @Override public fftw_iodim64_do_not_use_me position(int position) {
+    private native void allocateArray(long size);
+    @Override public fftw_iodim64_do_not_use_me position(long position) {
         return (fftw_iodim64_do_not_use_me)super.position(position);
     }
